refactor: Remove unused test-utils file
- Deleted the test-utils.tsx file as it was no longer needed after refactoring the test setup for the useTokenBalances hook.
- This change helps streamline the codebase and improve maintainability by eliminating redundant files.
